requirements.txt
pyautogen==0.8.3
autogen==0.8.3
openai==1.68.0
python-dotenv==1.0.1